readdir on InodeOrTreeOrEntry
Summary:
Buck2 needs to be able to list directory contents to search through the
filesystem for targets. It can use the filesystem for this, but on NFS
readdir can be suprisingly expensive, as many standard libraies implementation
of readdir will endup causing EdenFS to load all the children in the directory.

Here we add a method to our abstraction that avoids loading inodes when
unnessecary.

Reviewed By: chadaustin

Differential Revision: D36987096

fbshipit-source-id: 586da5544b46c471a480f4e7dd79211f239a3cc2