hg: testlib: Add a test debug mode to inspect server logs
Summary:
This diffs updates the testlib runner to include a debug mode that can
be enabled by setting `HGTEST_DEBUG`. This mode means that the temporary test
directory isn't cleaned up and Mononoke server's stderr is written to a file.

In the future, this debug mode could be expanded more (perhaps it could imply
that tracing should be enabled?).

Reviewed By: DurhamG

Differential Revision: D37137962

fbshipit-source-id: 2005c227ad17e099f22650d9d5eb78dc9322a742