add TreeOverlayTest
Summary: Adds a windows only `TreeOverlayTest` which tests the existing `TreeOverlay` class. It's based off of the `OverlayTest` with some modifications that take into account that files are not handled with this structure. It technically can run on Linux and macOS, but one test fails due to some windows only ifdefed code in the EdenMount initialization code.

Reviewed By: chadaustin

Differential Revision: D35629360

fbshipit-source-id: 20d6c5d89dcf6d390c4695742f729f715d0224ae