clone: add methods for getting bookmark names and values
Summary: Currently the revision to checkout is passed from the clone pull method to the checkout method. For the eden clone case, there may already be a backing repo cloned. Create a method to be shared between the sparse clone and eden clone for getting the checkout target revision from the metalog. Methods used for eden clone in D36911566

Differential Revision: D36913315

fbshipit-source-id: 06961137e6a4beae9cca9b2851724ae02e5bb6aa