Implement squash commit rewrite
Summary:
TLDR: Syncing repos with many merges in the history is problematic for us. If we preserve the history as is, we would need to rederive a lot of extra data. To overcome that we want to squash some of the branches into a single commit (mostly feature-like branches from a single author and up to some limit of commits). More context described here https://fb.quip.com/oDRQALiasyHb.
Next diffs will add the checks that all commits in the side branch are from same author and not exceed provided limit.
As of right now MergeMode::Squashed is not used, so it should be safe to land it.

Reviewed By: mitrandir77

Differential Revision: D37167202

fbshipit-source-id: 7ed81a87fba362ef2c14e82b8b7bd350a74a7efe