Initialize an ActivityBuffer per Eden Mount
Summary:
In order to enable retroactive debugging of past events in EdenFS, we have constructed and implemented a new data structure (called an ActivityBuffer) to store past events in a fixed size buffer that will be accessed by new retroactive versions of the Eden tracing CLI.

As the ActivityBuffer has finished being implemented, we now begin teaching EdenFS to use the ActivityBuffer. In this diff we start this process by initializing an ActivityBuffer for each Eden mount. Since we have a configuration option for whether we want to enable the buffer yet or not, we wrap the ActivityBuffer in an std::optional type and only store a value if the enableActivityBuffer configuration setting is true when the EdenMount is initialized.

Reviewed By: MichaelCuevas

Differential Revision: D37148921

fbshipit-source-id: ec1939011acf0dc3f51ccd6e8d430be856eaf077