inodes: default to Tree overlay in unittests on Windows
Summary:
The Legacy overlay errors out on Windows due to it being deprecated,
unfortunately there are a couple of places in tests that still default to it,
causing them to pass on Linux/macOS, but not on Windows. By defining a
kDefaultOverlayType and using it wherever the Legacy overlay was used, we can
avoid this issue from occuring in the future.

Reviewed By: chadaustin

Differential Revision: D37214271

fbshipit-source-id: 9217c6fd5a2b3bc7a4526db07a766f3e00f26aac