prjfs: fix race when listing directory in dispatcher
Summary:
There is a small race in between detecting that a directory exists with listing
it. If the directory is removed from the disk in between, the fileNotification
would fail with an error. Until recently, this error was merely logged, but is
now fatal in debug builds, thus failing tests. Let's properly handle it so
tests aren't flaky.

Reviewed By: genevievehelsel

Differential Revision: D37189377

fbshipit-source-id: 6305e5458e63c8b5998928c7cfc39c68949e6eae