utils: properly detect ENOENT on Windows
Summary:
On Windows, system error may be using the Win32ErrorCategory, and not the
generic system_category or generic_category. Thus we also need to test for that
category as well as the Windows error codes.

Reading a bit on the internet how Win32 error codes are supposed to be handled,
in theory, using the system_category should be the preferred method. However,
this would also mean that printing exceptions would also contain a \r at the
end, which feels suboptimal. Maybe xlog should just be fixed to never try to
print \r characters to avoid this issue?

Reviewed By: genevievehelsel

Differential Revision: D37189378

fbshipit-source-id: 03ca35c38a05bc35c8a184a6069b1ab41b6852c0