inodes: handle case change of a non-empty directory
Summary:
On Windows, changing the case of a directory is a bit finicky. EdenFS currently
tries to remove the content of the directory, then calls PrjDeleteFile on it
and finally add the new directory. In theory, this has the effect of removing
the directory name from ProjectedFS cache, and then adding the differently
cased name.

This works well (as the existing tests shows), until an untracked file is
present in the directory. In order to not lose user data, EdenFS will not
remove it, but still try to call PrjDeleteFile on the directory. Unfortunately,
PrjDeleteFile will not be successful on a non-empty directory, thus the case
change fails. Worse, EdenFS believes that directory is empty, causing
discrepency in `hg status`.

Testing this with Mercurial shows that Mercurial will change the case of
directories only when there are no untracked files, and as soon as an untracked
file is present, the directory case will be unchanged.

Thus to solve this, we can first try to change the case, and if that fails,
fallback to just updating the directory content.

Reviewed By: chadaustin

Differential Revision: D36332663

fbshipit-source-id: bfa20fb703142a2b6090fe2718400781cc18ef2e