Support setting bonsai/git mapping in SCS commit creation
Summary:
When we create commits that should have a known git hash, we want to record that mapping (remote gitimport, for example).

Make it possible to reuse the hggit extras for this automatically, gating this behind a permission check and a request for a git format hash.

Later, when we have the ability to derive git format commits, we can reuse this code path for derived hashes, too.

Reviewed By: markbt

Differential Revision: D37108292

fbshipit-source-id: 75256383ee5d8a9ff56689a4a5ebc9341fc86227