separate and rename EdenFsSample components
Summary:
EdenFsSample was a generic telemetry sample created for tracking CLI usage. The naming does not convey its ended use.

In addition, we should separate the scuba_builder logic out from the EdenFsSample impl so that we can use it to create other sample types in the future. An example of this is the PrefetchProfileSample which will be added in the next diff.

Reviewed By: fanzeyi

Differential Revision: D37011583

fbshipit-source-id: fe079642062499a96f0c63e834eaf0d97dba6139