service: collect streaming time of streamChangesSince
Summary:
The Thrift stats only expose the processing time of the Thrift handler, not of
the streaming part. In order to better understand the time actually spent in
streamChangeSince, let's collect data in it.

The code was written in such a way that we can easily expand it for all the
other Thrift requests, which can be done as a follow up change.

Reviewed By: chadaustin

Differential Revision: D37255869

fbshipit-source-id: 10b07b74d9516b58ab77bfc0a7eb3a650d13ec78