readdir endpoint
Summary:
Buck2 needs to be able to list directory contents to search through the
filesystem for targets. It can use the filesystem for this, but on NFS
readdir can be suprisingly expensive, as many standard libraies implementation
of readdir will endup causing EdenFS to load all the children in the directory.

We finally have all the plumbing we need, to answer a simple list of files
in a directory. We do not yet return any attributes about the files in the
directory.

Reviewed By: xavierd

Differential Revision: D36987124

fbshipit-source-id: 573bc0f84a157c51090e0fa3d3fdabad6b4ec79e