Move ActivityBuffer initialization directly into constructor
Summary: In diff (D37148921 (https://github.com/facebookexperimental/eden/commit/804d06d77a4427c2875c92ad260135b32a002a95)), the windows OSS build gave an error when initializing an ActivityBuffer for an EdenMount that it was attempting to reference a deleted function. This is a fix to move the ActivityBuffer initialization directly to the constructor of the ActivityBuffer so the windows OSS compiler doesn't try to do a copy after we already deleted the copy constructor for the ActivityBuffer.

Reviewed By: xavierd

Differential Revision: D37284065

fbshipit-source-id: ad9fea634e581e45df361153b1ef35242d40d5ef