status: fix deadlock in rust watchman workingcopy
Summary:
Integration tests exposed a deadlock: filechangedetector tries to lock treestate while state.merge was already holding onto a lock.

We fix this by splitting up both usecases into two lifetimes: merge step and a persist step. Funny how we end up back here but I think we've finally got this right now.

Also in this diff: state doesn't need to hold onto a filechangedetector for its entire lifetime and instead we pass it as an argument to merge. This allows us to be sure at the boundary there can be no double locking (and why I believe this two step API is the right approach in the end).

Reviewed By: DurhamG

Differential Revision: D37136050

fbshipit-source-id: 5a7980d066d88de30c15cc98e11814275ea88cf3