Delete --generate-bundles argument
Summary:
On D36702781 (https://github.com/facebookexperimental/eden/commit/d75d46b9740c438a59325a6e963aae568cedc18a) this argument was made to be ignored. This diff removes it from both the code (binary now crashes if it is provided, instead of ignoring it), and from the TW spec that uses it.

To be safe (in case of emergency reverts, which are unlikely), let's wait until the changes in D36702781 (https://github.com/facebookexperimental/eden/commit/d75d46b9740c438a59325a6e963aae568cedc18a) are deployed before landing this diff.

It's fine to land the TW changes at the same time as removing the argument, because TW always uses the most up to date specs, so it will always be ahead of the fbpkg with the removed argument.

Reviewed By: farnz

Differential Revision: D36934511

fbshipit-source-id: 2bc4a4a4e90cd061282d365a6376bd30f8c236c7