test: add RepoGenerator to GlobalTestState
Summary: Right now the file name pattern generated is the same for each repo: file0, file1, file2... If a repo pulls a commit with a file generated from another repo, the first filename it generates will be the same as a file that's already in the repo, instead of a new filename. Share the RepoGenerator as a test global so every new filename is unique.

Reviewed By: DurhamG

Differential Revision: D37218147

fbshipit-source-id: 252f933156dbb736f90bb0d8ca812594c548405f