store: split loadGitIgnoreThenDiffTrees
Summary:
One of the unfortunate behavior of this function is that it forces inodes to be
loaded when calling getLoadFileContentsFromPath. As we now know, loading inodes
is very expensive on some platforms and thus we should make sure to limit it.
In this case, we should be able to read the .gitignore file directly from the
ObjectStore without needing to load inodes. Splitting the function will help in
making this happen.

Reviewed By: kmancini

Differential Revision: D36363213

fbshipit-source-id: 6bb106cee4e77a33fc922c5e4f23215298f9d2cf