service: better support negative timeout in synchronizeWorkingCopy
Summary:
We've seen a couple of users with errors in their `eden rage` where
synchronizeWorkingCopy timed-out. Since this is only called in Watchman, which
passes a negative time for an infinite wait, this means that
`SemiFuture::within` misbehave with negative times.

Let's be explicit in these cases and not call `SemiFuture::within`

Reviewed By: chadaustin

Differential Revision: D37117518

fbshipit-source-id: 549e005850fb8925b667b23b75b3a088809a610e