Add Thrift Interface for accessing the ActivityBuffer
Summary:
In previous changes, we modified EdenFS to record InodeMaterialize events in a new ActivityBuffer data structure that was contained by each EdenMount. Now, in this diff we will provide a new endpoint to EdenFS’s thrift interface that will give access to these recorded InodeMaterialize events for a given mount point. With this endpoint, we will next create a eden trace inode —retroactive command to retrieve the recorded events and display them for the user.

Note, when creating the thrift data structures for this diff, I created a new InodeType thrift enum that I used to replace the previous enum in inodeUtils.

Reviewed By: kmancini

Differential Revision: D37288805

fbshipit-source-id: 1c8140b5bd56d38bb3b91c39bb840fdfd3b171b3