Don't save unbundle content to blobstore
Summary:
(I was wrong, the previous diff was not the last with stuff to delete.)

As suggested by farnz on D36934388 (https://github.com/facebookexperimental/eden/commit/22b65db292c2103fb2fff670f1757260e8b9de73), we don't need to calculate the bundle2 id anymore when closing the stream.

This diff makes sure we don't do that. It saves a write to blobstore.

The logic for keeping track of the request is still there, and I'll remove it in the next diff (D37003263).

Reviewed By: farnz

Differential Revision: D37002877

fbshipit-source-id: e2f1fa7a3641f5c6380c008035949bf92d0b3357