Delete bundle2_replay params and config
Summary:
This diff deletes the config related to "preserving raw bundle2 ids", which after this stack, is no longer necessary. All it was doing was writing the data to an `Arc<Mutext<Bytes>>` and never reading from it.

It also removes all config parsing related to that. The configerator counterpart is in D37003181.

Reviewed By: farnz

Differential Revision: D37003263

fbshipit-source-id: 3a639813877ec6f896ff9a0fe24020708e2ec798