tests/integration: use separate certs for client, proxy and server
Summary:
Currently the integration tests use a single certificate for the client, proxy and server, and that certificate is in the global trusted allowlist.  This means certain authorization-related features are not being tested, as they are always allowlisted and we have no way of knowing the difference between a permitted client and a non-permitted one.

Start to fix this by splitting out the certificates into three: one for a client (named `client0` to allow for expansion to further clients later on), one for any trusted proxy that might forward credentials (so we can have tests that ensure that only trusted proxies may do this), and one for the server itself.

Reviewed By: yancouto

Differential Revision: D36712641

fbshipit-source-id: 1b832b11213d406f2e21a457ab900eecf5c4b820