config: split security config into separate allowlists
Summary:
Currently users explicitly allowlisted are considered "trusted proxies" by
EdenAPI.  This has the side-effect of stripping their credentials from
the `CoreContext`: they are replaced by the headers from the request, as
we are expecting these requests to come from trusted proxies.

If entries in the allowlist are in fact users, this breaks authorization later
on for them, as we can no longer know their identity.

Separate out the allowlist for users that can always access the service (but
only as themselves), and the allowlist for services that are permitted to act
as trusted proxies (and so pass on authentication information that we trust).

Reviewed By: yancouto

Differential Revision: D36701824

fbshipit-source-id: 9a6d21fe7b4ebbffb6c3e47a539449e33058a1be