permission_checker: allow combination permission checkers
Summary:
Allow combinations of permissions checkers by expanding
`PermissionCheckerBuilder` to the full builder pattern.

When combined, permissions are unioned, so any grant of access is sufficient.

This will be used to have both a global allowlist and per-repo access control
(which may be provided by a remote ACL checking service), so ensure that
failure of a single checker does not cause failure overall: if one checker
fails, then we only return that error if all other checkers either fail or
return false.

Reviewed By: yancouto

Differential Revision: D36760752

fbshipit-source-id: bffd3dfbc892ab4daea4ce643f93235f86e31d49