mononokepeer: remove potentially large recursion
Summary:
When reading a segment off the wire, if mononokepeer reads a stderr
message, it uses recursion to try to read the next segment. If there are a lot
of stderr messages in a row, this could lead to a large stack.

Let's use a while loop instead, to avoid the stack bloat.

Reviewed By: jordanwebster

Differential Revision: D37360504

fbshipit-source-id: 66bf4ff8c729ecada7d4e72968a1b49a7b591ad3