whereami: implement as modern Rust command
Summary:
Currently whereami is implemented in the Rust telemetry wrapper. Now that we have Rust command dispatch technology, move it in to hg proper.

The telemetry wrapper's semantics diverged slightly from the original Python extension. I went with the Python's behavior of printing out an all zeros hash if there is no dirstate.

Reviewed By: DurhamG

Differential Revision: D37348809

fbshipit-source-id: 48bbf950e8cba96920c5fe382b7ff8f7cce79be2