dirstate: implement basic deserialization
Summary:
Implement Rust deserialization of the dirstate file (which is just a pointer to the treestate file).

I made it return a partial dirstate object even if it isn't a treestate dirstate. This is so we can at least read the parents for other types of dirstates (e.g. EdenFS).

Reviewed By: DurhamG

Differential Revision: D37348803

fbshipit-source-id: b191034b49050f4746ccd44fc885508ee75b49cb