kill process with open files before removing mount
Summary:
When a process still has an open file in an eden redirection when we try to
remove it, there seem to be problems.

On linux we can unmount with the lazy option which allows removing the mount
from the filesystem, but only destroying it when processes no longer hold open
files. On mac there is no such option, so we just force unmount. On CI sometimes
we see that we are not able to mount volumes for mysterious reasons. These seem
to be correlated with us trying to unmount redirections where processes still
have open files.

As a mitigation we will try killing processes with open files before we unmount
them in the apfs mount helper.

Reviewed By: Artemmkin

Differential Revision: D37164579

fbshipit-source-id: f72f90b6196292f42a079891018fd3117beededf