Allow forwarding identities from trusted sources
Summary:
Just like already happens for proxygen on Mononoke, we want to be able to forward identities from Mononoke to SCS calls (and will need to do this more and more as we add "internal services").

This diff adds code for that both on client (Mononoke) and server (SCS). On SCS, I add the ability to accept forwarded identities if the client is trusted (passes hipster tier check). On Mononoke, I add the code to provide such headers to forward identities by default on the pushrebase calls.

Reviewed By: markbt

Differential Revision: D37344951

fbshipit-source-id: 0289d481824e6b6ff1aa7f26dce7b1653fe6413c