hgcommands: re-organize some code
Summary:
Separate the command dispatch logic out to a separate function. My main goal was to separate the cant-create-dispatcher error from other command errors. The error handling for the two were mixed together, which resulted in a lowest common denominator situation where the config and other essentials were not available.

Now a dispatcher creation error is handled before the dispatch logic. This change assumes that Dispatcher::from_args() can't itself return the FallbackToPython error, which I think is the case.

Reviewed By: DurhamG

Differential Revision: D37381853

fbshipit-source-id: 048e8ae4b50a9668db919cdacde9f275927bb473