Sharded map should not depend on DMv2
Summary:
This diff makes none of the `sharded_map` code depend on anything specific to DMv2.

The last bit that was depending was that sharded map node ids were always serialized with `deletedmanifest2.mapnode.XXX` blobstore keys.

This diff makes that customisable, by allowing you to specify the `Id` that will be used when using sharded map, and dmv2 specifying the dmv2 specific one. It also renames the id structs to make that clear. I had to add a new `IdContext` trait for it, which allows us to use Context structs in a generic way.

This allows other stuff to use sharded maps with their own hashes.

Reviewed By: mitrandir77

Differential Revision: D37112685

fbshipit-source-id: 12b22f33ee4b43cc1ff06eb44152adabac33366c