git: fix commit --amend
Summary:
When doing 'hg commit --amend', prepping the diff stats for the commit
message template requires computing memctx.status(). This requires computing
in-memory hashes for all the changes. The old code was hardcoded for Mercurial,
and required a filelog that could answer parents(). Git doesn't have that, and
doesn't need that data to compute the hash, so let's special case it.

Reviewed By: quark-zju

Differential Revision: D37357793

fbshipit-source-id: d31230cb1dd6f66663bf84d1dcd501a9bad1a66a