Change Inode diff to accept a vector of trees
Summary:
In a future diff we'll need to check if the inode is different from
both tree A and tree B. To do so, we'll need to pass a vector of trees to the
diff algorithm. For now, let's just update the function signatures, even though
it only processes a single tree still.

Reviewed By: xavierd

Differential Revision: D36914466

fbshipit-source-id: cc9a9c2b1173afdf7fd008b0c56f8a7c2512640c