Expose the original checkout source and dest on ParentCommitState
Summary:
An upcoming diff will allow resuming an interrupted checkout. To do so,
we need to be able to read the stored source/dest of the interrupted checkout.
This is stored on the ParentCommit. This diff reads those values and stores them
on the ParentCommitState so they are available at checkout time.

Reviewed By: xavierd

Differential Revision: D36914463

fbshipit-source-id: 1a94e0a5cc9b334e69766b76296b0b869f6cf039