Allow resuming an interrupted eden checkout
Summary:
Previously if the user killed an Eden checkout mid-operation, we would
force them to reclone. This was because the root inode no longer matched the
current parent commit, which caused Eden to think a large portion of the working
copy was dirty, and try to materialize it.

The snapshot file already tracked the from/to commits, so this diff just changes
EdenFS to treat the "from" commit as the parent commit, instead of the to. In
other words, edenfs doesn't consider itself as on the new commit until the
commit actually finishes now.

Reviewed By: xavierd

Differential Revision: D36914461

fbshipit-source-id: 3dd84deb655114c12d637d8e761aaf62f95fe44f