Don't change state during dryrun checkout
Summary:
In a future diff we'll support resuming an interrupted to checkout. To
do so, we need to make sure a dryrun checkout doesn't overwrite the interrupted
checkout state. Notably, we need to not change the in-memory value of
checkoutInProgress, and we need to not change the mount's checkout config.

The checkoutInProgress bit wasn't an issue before because we didn't allow a
checkout when one was in progress. But soon we will.

I'm unsure of the impact of the mount's checkout config changing during a
dryrun. Perhaps this was fine?

Reviewed By: xavierd

Differential Revision: D36914465

fbshipit-source-id: d20c2c936c3ebeafcf4fdf3d6c131f3d2d64991a