Print warning message in eden trace inode --retroactive to enable activitybuffer
Summary: In a previous change (D37356739 (https://github.com/facebookexperimental/eden/commit/d5573cf20bfc6575a66ea09117440efada18d5f3)), we added support for the `eden trace inode --retroactive` subcommand. However, in order for this command to work, the current EdenMount must have an ActivityBuffer initialized. Previously, if no ActivityBuffer was initialized this command would simply give an error. In this change, we make sure to catch this error explicitly and provide a message to the user suggesting that they should enable the ActivityBuffer

Reviewed By: MichaelCuevas

Differential Revision: D37464704

fbshipit-source-id: 8b93ed5eea63da8d15b7d73a7fc6e5fa98f110ee