Create new command, using new command lib, and move cat
Summary:
This creates a new scsc command using the new command lib.

I hg copied all the relevant files so that the history is preserved nicely, as I plan to delete the old code once I move all subcommands.

The new command has a new target and can be used in integration tests. I will not deploy it or add it to fbpkgs, as the plan is to quickly replace the original scsc with this new implementation, and the copy is just kept in place so that this can be landed diff by diff.

Reviewed By: kris1319

Differential Revision: D37421537

fbshipit-source-id: 8ca13bc95d500d5a216d76e021ef463dfdb657c5