Create base_app library
Summary:
This diff separates the base logic of the `mononoke_app::subcommands` macro to a different crate, and uses it as a helper so all use cases today continue working.

This enables us to use something other than `MononokeApp` with the macro, which is the plan for  `scsc` on the following diffs.

Reviewed By: kris1319

Differential Revision: D37458912

fbshipit-source-id: 6df05988234f1741c6250ab404bd6838662270f4