wal multiplexedblob: implement reads
Summary:
Reads on multiplexed blobstore:
- return blob as soon as any of the underlying `get`s successfully returns blob
- return `None` when at least read-quorum blobstores returned `None` (we don't wait on the other `get`s)
- fail if none of the blobstores could find the blob and no read-quorum on `None` was achieved.

It is possible for multiplexed `get` to return inconsistent results if the previous `put` failed. For example, the `put` operation successfully logged itself to the WAL, but 2 out of 3 underlying `put`s failed. It means only 1 blobstore has the blob.

Now multiplexed `get` tries to read from the blobstores: first 2 underlying `get`s returned `None` -> quorum achieved, multiplexed returns `None`. But if first returns the blobstore that actually has the blob, multiplexed `get` will return blob.

This is fine because the only guarantee we have: if the `put` succeeded, the blob will be readable from the storage. But in case like above, the blob will be either eventually replicated to all the blobstores by healer or will be rewritten by another `put` (because the first `put` failed).

Reviewed By: farnz

Differential Revision: D37338032

fbshipit-source-id: 6148075dfcce6032628565249f54f742065baf1d