wal multiplexedblob: implement is_present
Summary:
Is_present on multiplexed blobstore:
- return `Present` as soon as any of the underlying `is_present` successfully returns `Present`
- return `Absent` when at least read-quorum blobstores returned `Absent` (we don't wait on the other `is_present`s)
- return `ProbablyNotPresent` if some blobstores returned `Absent` (but not enough for quorum), the rest failed
- fail if all of the blobstores failed.

It is possible for multiplexed `is_present` to return inconsistent results if the previous put failed. For example, the put operation successfully logged itself to the WAL, but 2 out of 3 underlying puts failed. It means only 1 blobstore has the blob.

Now multiplexed `is_present` tries to read from the blobstores: first 2 underlying `is_present` returned `Absent` -> quorum achieved, multiplexed returns `Absent`. But if first returns the blobstore that actually has the blob, multiplexed `is_present` will return `Present`.

This is fine because the only guarantee we have: if the put succeeded, the blob will be readable from the storage. But in case like above, the blob will be either eventually replicated to all the blobstores by healer or will be rewritten by another put (because the first put failed).

Reviewed By: farnz

Differential Revision: D37342528

fbshipit-source-id: 21c62a9ecd2c2d117ee0b73d073f6ecd03f383c1