integration: fix test_change_casing_with_untracked on Windows
Summary:
In D36914467 (https://github.com/facebookexperimental/eden/commit/e7918c1d11f1158648a371c7699aa8fd7d4eefae), a very subtle change was introduced in the processBothChanged
lambda. The path component being used before was the one from the tree, while
it is now coming from the inode itself. This change caused the
test_change_casing_with_untracked to start failing with:

  AssertionError: {'DIR2/untracked': '?'} != {'dir2/untracked': '?'}
  - {'DIR2/untracked': '?'}
  ?   ^^^

  + {'dir2/untracked': '?'}
  ?   ^^^

Or in english, the test expected the untracked file to be in the directory
DIR2, but is now in the directory dir2. Reading the test itself, and the
comment just above the failed assert makes it clear that on Windows, the dir2
directory will be on disk. It thus makes more sense for status to report the
untracked file from the dir2 directory instead of from the DIR2 directory.

Reviewed By: DurhamG

Differential Revision: D37536084

fbshipit-source-id: 02dd592719f41ad0320f065ccfe41dc64d030608