service: move diff computation to the background
Summary:
The main benefit of the streaming API is that a client can stop it by simply
closing the connection and EdenFS will detect this and stop the diff.
Unfortunately, one drawback (or benefit, depending on what is desired) of
ImmediateFuture is that they execute eagerly as long as data is available and
can be computed.

The negative side effect is that the streaming API may diff the entire trees if
these are all in memory or present locally without giving a chance to the
Thrift client to stop it. To prevent this, we can just make sure that the
computation is done lazily by forcing the computation to be done in a
SemiFuture.

Reviewed By: quark-zju

Differential Revision: D37539114

fbshipit-source-id: b69400ea6f7c3d7c3789fd0d0ab164c49a74aa19