Fix identity parsing from string
Summary:
The code assumed that the identity never had any `:` in it (and checked for that and failed in that case, at least).

But that is not true, I can see some `MACHINE:hostname` identities where hostname is ipv6 so has some `:` in it, which causes failures on remote pushrebase (I've only seen 2 yesterday though, it's not that common).

It's not a blocker for the migration because we have local fallback and it's rare.

Reviewed By: kris1319

Differential Revision: D37517177

fbshipit-source-id: 7e4181ab921e5817fbf70c5123e9b50088f8535d