mononoke: hg_sync: Double write with new RepoLock
Summary:
Update Mononoke code to use the new `RepoLock` trait to double write
to both `repo_lock` DBs. As Mononoke only locks or unlocks repos from the
hg_sync_job most of the changes are there.

In a later diff, I will remove the code that writes to the old hgsql DB.

Reviewed By: mitrandir77

Differential Revision: D36129277

fbshipit-source-id: 04b173d70213da8bfb79ead55980f033c7453927