Use AsyncManifest in derive
Summary:
This uses `AsyncManifest` in the common manifest derive method.

This will unblock us to start writing manifest with sharded maps.

It still does not reap its benefits, though, all code is sequential and traverses all subentries unnecessarily, and will need a decent refactor in order to reap the benefits of sharded maps.

Reviewed By: mitrandir77

Differential Revision: D37174806

fbshipit-source-id: 0f98140f659effb9703837a71be668920136a9d0