Add ThriftConvert to MononokeId and implement it to sharded map ids
Summary:
This makes ids more generic, by moving stuff out of the macros and into trait implementations, like the `Storable` definition.

This allows us to use the "default thing" to create ids for the sharded map nodes stuff, which before used some custom stuff. This is better and cleaner. It allows us to remove some duplicated code (loading/storing in blobstore) from sharded maps as well.

It was a pain in the ass to make the diff, because the types are very intertwined with the macros, it's hard to decouple.

Reviewed By: kris1319

Differential Revision: D37139434

fbshipit-source-id: 92507a67bdfb42652fdd05d3bcba66ba44c13752