store: only manually reset ThreadLocalPtr on Windows
Summary:
krallin saw an ASAN failure when running tests pointing to a double free
happening when destructing the HgImporterManager. The ASAN backtrace looked
weird mostly due to the HgImporterManager seemingly being freed in 2 different
threads, which is nonsensical. The hypothesis is that the per-thread call to
`reset` and a global cleanup both race to destruct it, causing what we're
seeing.

The solution in this case is fairly straightforward, since on Linux/macOS the
ThreadLocalPtr will be correctly destructed when the thread is destroyed, we
can simply omit the SCOPE_EXIT there, and only have it on Windows. This'll
avoid the double free.

Reviewed By: kmancini

Differential Revision: D37215356

fbshipit-source-id: 92c15af7348652befb1e7ea9cfa9dac23b672269