prjfs: fix use after free in handleNotPresentFileNotification
Summary:
In order to avoid copying paths and allocation, this code cleverly uses path
piece on a path that is moved and will live for longer than the path pieces.
Unfortunately, the assumption on which this optimization was built is not
always true. To understand more, we need to dig onto the implementation details
of RelativePath and of std::string.

A RelativePath is merely a wrapper around an std::string, and a
RelativePathPiece is also just a wrapper on top of a folly::StringPiece (which
has similar semantic as std::string_view), that is, a RelativePathPiece merely
holds a view of to the underlying std::string buffer. When a RelativePath is
being moved, the held std::string is also moved, which in most cases prevents
copying and re-allocating a string, this makes the move operation fairly cheap
and since the RelativePathPiece were a view on that first string memory
allocation, these are still viewing valid and allocated memory.

However, std::string have an optimization where small strings aren't heap
allocated, but are stored in the std::string object itself, this is called SSO
for small string optimization. In this case, a folly::StringPiece is no longer
a view on the heap allocated memory, but on that SSO memory. What this means is
that moving a SSO std::string will make the folly::StringPiece invalid as it
would no longer point to valid memory!

As mentioned above, RelativePath and RelativePathPiece are just wrappers on top
of std::string and folly::StringView, and thus the same caveat exist. This
means that the optimization performed in handleNotPresentFileNotification is
correct only for RelativePath long enough to not trigger the SSO.

Since this pitfall is extremely easy to fall into, follow up changes will
attempt to reduce the likelyhood of falling into them.

Reviewed By: genevievehelsel

Differential Revision: D37422971

fbshipit-source-id: 610bcf955d53498710b79f12754cd2b1344bc001