service: increase loglevel for Thrift request longer than 1s
Summary:
In a previous diff, I raised the loglevel of streamChangesSince, but that made
it too verbose, due to Watchman calling it very frequently. Since the goal was
to make it visible when long requests are occuring, we can instead raise the
log level for long requests. For now this is defined as any request over a
second.

Reviewed By: genevievehelsel

Differential Revision: D37652964

fbshipit-source-id: 646cb75b797abcbfd7c3a7d50b69b5a9371abe7b