Enabling repo-context in logging by including repo-name
Summary:
The logs by-default do not contain the repo-name when getting output on the console. In case of sharded jobs, multiple repos could be executing on the same host and the logs from different repos might overlap. Without having a repo-identifier, it is very difficult to figure out the source repo while debugging using the logs. This diff adds the necessary repo-context in the form of repo-name to the `info`, `warn` and `error` logs generated by Mononoke.
Additionally, this diff includes changes to the integration tests where the output would now be different due to the repo-name being included in the logs.

Reviewed By: markbt

Differential Revision: D37646225

fbshipit-source-id: 5c1cf7bc5fb8e81e990a25dcccd2ae822dd6f9ca