store: reset threadLocalImporter when importer thread terminate
Summary:
There is a subtle bug in the code currently due to the `folly::kIsWindows`
condition creating a scope, and thus the `SCOPE_EXIT` will execute right
after the condition scope ends. This occurs when the thread starts, but the
intention is for that to happen when the thread terminate.

Moving the `folly::kIsWindows` check in the `SCOPE_EXIT` solves this.

Reviewed By: fanzeyi

Differential Revision: D37673272

fbshipit-source-id: 4cd099bd4cf3b3d931abd1cb267308cfa5190cd3