mononoke_api: merge RepoDraftContext into RepoContext
Summary:
To simplify mononoke_api, don't require users to get a `RepoDraftContext` for
draft changes.  Instead, checks for whether draft changes are permitted are
deferred to later on using the `AuthorizationContext` of the user, as depending
on configuration it may depend on the parameters.

One limitation here is that we still need to check the repo config to see if
source control service writes are enabled at all.  In the future we should roll
this config option into the general repo read-only state, but for now we just
remember to check it manually when we start the draft operation.

Reviewed By: HarveyHunt

Differential Revision: D36552807

fbshipit-source-id: 5028f9a4a8180ac63e92b85f7a676e3fde543534