scs_server: check megarepo permissions via AuthorizationContext
Summary:
The megarepo tooling currently checks if the user has permission to write to the repo by the side-effect of creating a `RepoWriteContext`.  That's not ideal, because it doesn't actually *use* the `RepoWriteContext`, so this could easily get missed or lost.

Instead, change to a manual checking of permissions using the `AuthorizationContext`.

We also need to check that writes are enabled, so (as on the earlier diffs) we must do this manually, too.

Reviewed By: yancouto

Differential Revision: D37500875

fbshipit-source-id: e2b0791bf7431272224e9c100215282073a481d8