Use channel permits to close the race condition in `GitReader`
Summary:
S277945 is caused by a race in `GitReader`. It is possible for an import process to call `get_object`, have the object that's been asked for by a previous call returned to it, but then never complete the `send_request.send(oid).await?` to send the request to the task that writes to `git cat-file`'s `stdin`.

Use permits to close the race - we first get a guarantee that we can send our request, then we add ourselves to the interest list for this object, then we send the request (which cannot be delayed).

Reviewed By: mitrandir77

Differential Revision: D37688657

fbshipit-source-id: ae716743e7cfcb614145bcbc4502ea674497e537