TreeInode: get InodeOrTreeOrEntry for each entry
Summary:
We want to be able to expose attribute information for all the entries in a
directory with the readdir endpoint. This will allow Buck2 to access attributes
it needs while listing directories (filetype) and experiment with fetching all
the data it ever needs at once.

The first step towards this is fetching all the child InodeOrTreeOrEntrys for
a directory.

Reviewed By: chadaustin

Differential Revision: D36987150

fbshipit-source-id: 3b3191247b257c02504b554edaa7fc0bf18255b0