hgevents: don't eat errors during watchman client creation
Summary:
S278910 was caused by hg no longer instantiating watchman. This was
caused by a type error during watchmanclient creation. This should've been
surfaced instead of being silently eaten. Let's remove this catch.

If we need to, we can come back later and add a more narrow catch if there are
exceptions that should be eaten.

Reviewed By: sggutier

Differential Revision: D37721672

fbshipit-source-id: ff4e2bfb6b13ead356b71e57e68fba601388f432