mononoke: Replace read_write_fetcher with RepoLock facet
Summary:
Now that we have been double writing to `repo_lock` DBs for a while,
the old code can be removed.

This diff removes the code as well as the parsing for the
`write_lock_db_address` config option. I'll remove the config in a later diff.

As `RepoLock` is a facet, this diff also does some refactoring to make code a
bit clearer.

The diff also adds a new repo lock checker that can be used from within a
transaction. This is required as the pushrebase flow starts a transaction to
the metadata DB.

Reviewed By: markbt

Differential Revision: D36129275

fbshipit-source-id: dcfb9f06e76b21dbda74b5de4f9a359ebd32fff6